requirements.txt
Pillow==11.2.1
torch==2.7.0
transformers==4.52.3
torchvision==0.22.0

